Apps refactored. Todo app removed and todo functionalities implemented into users app. All imports are made absolute. New migration is created for Todo model
