--- server/apps/todo/admin.py
@@ -1,4 +1,0 @@
-from django.contrib import admin
-from . import models
-
-admin.site.register(models.Todo)
