--- server/apps/todo/serializer.py
@@ -1,8 +1,0 @@
-from rest_framework import serializers
-from . import models
-
-
-class TodoSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = models.Todo
-        fields = "__all__"
