+++ server/apps/urls.py
@@ -6,5 +6,4 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     # Complete authentication using custom JWT token
     path('api/', include('apps.users.urls')),
-    path('todo/', include('apps.todo.urls')),
 ]
